Fixed issues that occurred at compile time in macOS

+++ server/channel_manager.cpp
@@ -162,7 +162,7 @@ std::shared_ptr<Channel> ChannelManager::getChannel(std::string channel_id)
     }
 
     // Create the channel object and return it
-    channel = std::make_shared<Channel>((std::string)channel_res[0], (bool)channel_res[1], channel_res[2]);
+    channel = std::make_shared<Channel>((std::string)channel_res[0], (bool)channel_res[1], (int)channel_res[2]);
     return channel;
 }
